Experiencia de Aprendizaje 2 - Desarrollo de Agente Inteligentes con AI/Clase 2.4/requirements.txt
langchain-core
langchain-openai
langgraph
openai
plotly
streamlit
python-dotenv
